--- src/main/java/kitchenpos/products/domain/ProductProfanityCheckClient.java
@@ -1,5 +1,0 @@
-package kitchenpos.products.domain;
-
-public interface ProductProfanityCheckClient {
-    boolean containsProfanity(String text);
-}
