+++ src/main/java/kitchenpos/profanity/infra/ProfanityCheckClient.java
@@ -1,0 +1,5 @@
+package kitchenpos.profanity.infra;
+
+public interface ProfanityCheckClient {
+    boolean containsProfanity(String text);
+}
