+++ src/test/java/kitchenpos/products/domain/ProductNameTest.java
@@ -3,6 +3,8 @@ package kitchenpos.products.domain;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import kitchenpos.products.exception.InvalidProductNameException;
+import kitchenpos.profanity.infra.FakeProfanityCheckClient;
+import kitchenpos.profanity.infra.ProfanityCheckClient;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -11,17 +13,17 @@ import org.junit.jupiter.params.provider.ValueSource;
 
 class ProductNameTest {
 
-    private ProductProfanityCheckClient productProfanityCheckClient;
+    private ProfanityCheckClient profanityCheckClient;
 
     @BeforeEach
     void setUp() {
-        productProfanityCheckClient = new FakeProductProfanityCheckClient();
+        profanityCheckClient = new FakeProfanityCheckClient();
     }
 
     @DisplayName("상품 이름은 Null 일 수 없다.")
     @Test
     void nullException() {
-        assertThatThrownBy(() -> new ProductName(null, productProfanityCheckClient))
+        assertThatThrownBy(() -> new ProductName(null, profanityCheckClient))
             .isExactlyInstanceOf(InvalidProductNameException.class);
     }
 
@@ -29,7 +31,7 @@ class ProductNameTest {
     @ValueSource(strings = {"비속어", "욕설"})
     @ParameterizedTest
     void profanityException(String profanity) {
-        assertThatThrownBy(() -> new ProductName(profanity, productProfanityCheckClient))
+        assertThatThrownBy(() -> new ProductName(profanity, profanityCheckClient))
             .isExactlyInstanceOf(InvalidProductNameException.class);
     }
 }
